Add コメント取得処理を独立にする
