+++ src/app/api/comments/route.ts
@@ -3,23 +3,6 @@ import { NextResponse } from 'next/server'
 import { Comment } from '@/features/comments/types'
 import { prisma, main } from '@/lib/prisma'
 
-export async function GET() {
-  try {
-    await main()
-
-    const comments = (await prisma.comment.findMany({
-      orderBy: {
-        createdAt: 'desc',
-      },
-    })) as Comment[]
-    return NextResponse.json({ message: 'Success', comments }, { status: 200 })
-  } catch (err) {
-    return NextResponse.json({ message: 'Error', err }, { status: 500 })
-  } finally {
-    await prisma.$disconnect()
-  }
-}
-
 export async function POST(req: Request) {
   try {
     const { userId, userName, pictureId, body } = await req.json()
