+++ src/features/comments/components/Comment.tsx
@@ -17,20 +17,16 @@ import CommentCount from './CommentCount'
 import CommentForm from './CommentForm'
 import CommentsList from './CommentsList'
 
-import type { Comment as CommentType } from '../types'
 import type { AuthUser } from '@/features/auth/types'
 
 type CommentProps = {
   pictureId: string
-  comments: CommentType[]
   user: AuthUser
 }
 
-const Comment = memo(({ pictureId, comments, user }: CommentProps) => {
-  const { isLoading, onSubmitComment, handleDeleteComment, handleUpdateComment } = useMutateComment(
-    pictureId,
-    user,
-  )
+const Comment = memo(({ pictureId, user }: CommentProps) => {
+  const { isLoading, onSubmitComment, handleDeleteComment, comments, handleUpdateComment } =
+    useMutateComment(pictureId, user)
   return (
     <div>
       <Dialog>
